refactor(frontend): restructure partner pages into subdirectory layout
- Move partner pages into partner/{register,dashboard,payments}/index.tsx
  and shared components into partner/components/ to match the pattern
  used by booking/, property/, and search/
- Add partnerId to AuthUser type and propagate it through MFA login
- Dashboard and payments pages now use user.partnerId (not user.id)
  as the API identifier
