+++ frontend/src/pages/partner/components/MetricCard.tsx
